EDIT GameState/PlayerStart/PlayerState/PlayerCharacter to handle the third team spawning behavior.
EDIT Teleport Volume to handle teleporting to other parts of the map. Had to adjust the L_Hallway map to make it seem more seamless. NOTE: flashing lights should either be made steady, or reduce the brightness.

EDIT L_Hallway to include third area for team C.

BUG every so often the players will spawn a top of each other in the lobby